+++ app/src/main/java/com/example/payment/userDb/UserRepository.kt
@@ -3,7 +3,7 @@ package com.example.payment.userDb
 import androidx.lifecycle.LiveData
 
 class UserRepository(private val userDao: UserDao) {
-    val userBudget : LiveData<Float> = userDao.getUserBudget()
+    val userBudget : LiveData<Int> = userDao.getUserBudget()
     val userDetails : LiveData<User> = userDao.getUserDetails()
 
     suspend fun insertUser(user: User){
